refactor: FRONTEND: keep separating singletons into components/singles
