+++ frontend/src/components/ConsoleLogs.jsx
@@ -1,0 +1,9 @@
+const ConsoleLogs = ({logs}) => {
+  return (
+    <ul className='logs-list'>
+      {/* {logs.map((l, i) => <Log key={i} log={l}/>)} */}
+    </ul>
+  )
+}
+
+export default ConsoleLogs
