+++ frontend/src/components/Errors.jsx
@@ -1,0 +1,9 @@
+const Errors = ({errors}) => {
+  return (
+    <ul className='errors-list'>
+      {/* {errors.map((e, i) => <Error key={i} error={e}/>)} */}
+    </ul>
+  )
+}
+
+export default Errors
