+++ frontend/src/components/singles/Error.jsx
@@ -1,0 +1,11 @@
+const Error = ({error}) => {
+  const {content, line, payload} = error
+
+  return (
+    <li className='error'>
+      {`${content.slice(0, 41)} - ${JSON.stringify(payload)} - Line: ${line}`}
+    </li>
+  )
+}
+
+export default Error
