+++ frontend/src/components/singles/ExtraInfoSearch.jsx
@@ -1,0 +1,14 @@
+const ExtraInfoSearch = ({searchExtraInfo}) => {
+  return (
+    <>
+      <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-search"><circle cx="11" cy="11" r="8"></circle><line x1="21" y1="21" x2="16.65" y2="16.65"></line></svg>
+      <input 
+        type="text" 
+        placeholder="Search request by name"
+        onChange={(e) => searchExtraInfo(e.target.value)}
+      ></input>
+    </>
+  )
+}
+
+export default ExtraInfoSearch
